Remove NumberPool and ChatRoom
The Conduct Hotline had concepts of ChatRooms and Relays to obfuscate
the phone number of a reporter when contacting the hotline via
SMS. Nexmo numbers could be rented for one of two purposes --
either to be used as an event number, or as an SMS relay number.
Remove all of this functionality and associated models, they're
not relevant.